backend/functions/notifier/requirements.txt
firebase-admin==6.1.0
google-cloud-secret-manager==2.16.1
Jinja2==3.0.3
python-dateutil==2.8.1
Werkzeug
qrcode==7.4.2